+++ packages/auth/tests/OAuthTest.php
@@ -26,6 +26,7 @@ use Tempest\Auth\OAuth\OAuthClientInitializer;
 use Tempest\Auth\OAuth\OAuthUser;
 use Tempest\Container\Container;
 use Tempest\Container\GenericContainer;
+use Tempest\Mapper\MapperCache;
 use Tempest\Mapper\MapperConfig;
 use Tempest\Mapper\Mappers\ArrayToObjectMapper;
 use Tempest\Mapper\ObjectFactory;
@@ -37,7 +38,7 @@ final class OAuthTest extends TestCase
     }
 
     private ObjectFactory $factory {
-        get => $this->factory ??= new ObjectFactory(new MapperConfig([ArrayToObjectMapper::class]), $this->container);
+        get => $this->factory ??= new ObjectFactory(new MapperConfig([ArrayToObjectMapper::class]), $this->container, new MapperCache());
     }
 
     #[Before]